// src/api-generated/model/alocacao-horario-response-dto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * API Horários Acadêmicos
 * Documentação da API para o Sistema de Elaboração de Horário e Atribuição de Disciplinas
 * OpenAPI spec version: 1.0
 */
import type { AlocacaoHorarioResponseDtoDiaDaSemana } from "./alocacao-horario-response-dto-dia-da-semana"
import type { TurmaBasicaDto } from "./turma-basica-dto"

export interface AlocacaoHorarioResponseDto {
  /** ID único da alocação */
  id: string
  /** ID da turma alocada */
  idTurma: string
  /** Dia da semana da alocação */
  diaDaSemana: AlocacaoHorarioResponseDtoDiaDaSemana
  /** Hora de início da aula */
  horaInicio: string
  /** Hora de fim da aula */
  horaFim: string
  /** Data de criação da alocação */
  dataCriacao: string
  /** Data da última atualização */
  dataAtualizacao: string
  /** Dados da turma alocada */
  turma: TurmaBasicaDto
}