// src/api-generated/model/alocacao-horario-response-dto-dia-da-semana.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * API Horários Acadêmicos
 * Documentação da API para o Sistema de Elaboração de Horário e Atribuição de Disciplinas
 * OpenAPI spec version: 1.0
 */

/**
 * Dia da semana da alocação
 */
export type AlocacaoHorarioResponseDtoDiaDaSemana = typeof AlocacaoHorarioResponseDtoDiaDaSemana[keyof typeof AlocacaoHorarioResponseDtoDiaDaSemana];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AlocacaoHorarioResponseDtoDiaDaSemana = {
  SEGUNDA: 'SEGUNDA',
  TERCA: 'TERCA',
  QUARTA: 'QUARTA',
  QUINTA: 'QUINTA',
  SEXTA: 'SEXTA',
  SABADO: 'SABADO',
} as const;